requirements.txt
Flask==3.0.3
tensorflow==2.15.0
numpy==1.24.3
Pillow==10.0.0
gunicorn==21.2.0